Remove Plot module
Modules tend to spread throughout the code base (a refactoring I
attempted ended up with everything in a module) and are annoying to work
with. They make it hard to reason about order of
evaluation (particularly with cyclic dependencies) and difficult to
extend the code from the outside.

This refactoring replaces the module dependency with a given dependency.
This achieves the same effect, and is transparent to the end user, at
the cost of some possibility of difficult to debug compilation errors if
the user doesn't import everything correctly.