Sketch theming text
This commit adds theming for text at a Plot wide level. There is
a separate font for headings and normal text, but otherwise theming is
shared between all text.

Annotations are not currently themed.

I think this approach will very hit quickly hit a limitation. For
example, we ideally would separate theming for title, axis text, legend
text, and other text. For the title we want to be able to specify space
below it. So there are a lot of options necessary here. This simple
minded approach won't scale very well. It's a start, however.