Don't use empty to calculate bounding box
The `BoundingBox` from Doodle does not the semantics we want for
calculating the bounding box of data. `BoundingBox` must always contain
the origin, which is not the case for data. `BoundingBox.empty` is a
point at the origin. Using it means we include the origin when
calculating the bounding box of data, which is incorrect.

There must be a least one `Layer` in a `Plot`, so use a `NonEmptySeq` to
represent this.